A bug in CMOC caused files to not be closed. Tape files were not working. I've fixed the issue and added a tape read/write example.

+++ src/lib/unistd/close.c
@@ -3,12 +3,12 @@
 
 int close(int fd)
 {
-    if (fd <= _fcbact) {
+    if (fd <= (int)_fcbact) {
         return systemf("CLOSE#%d", fd) ? EOF : 0;
     }
     if (fd > FD_MAX) {
         _close((struct FileDesc*)fd);
-        free((void*)fd); // We are responsible for freeing fd
+        free((void*)fd);                            // We are responsible for freeing fd
     }
     return 0;
 }
